+++ rust/dbn/src/decode/dbn.rs
@@ -7,6 +7,8 @@ pub fn starts_with_prefix(bytes: &[u8]) -> bool {
     bytes.len() > DBN_PREFIX_LEN && &bytes[..DBN_PREFIX_LEN] == DBN_PREFIX
 }
 
+#[doc(hidden)]
+pub mod aligned_buffer;
 mod sync;
 pub(crate) use sync::decode_iso8601;
 pub use sync::{Decoder, MetadataDecoder, RecordDecoder};